IDM-304 "Introducing back the error checking code that had got removed. This had introduced a bug."

+++ Source/springidm/src/main/java/com/rackspace/idm/domain/service/impl/DefaultClientService.java
@@ -377,6 +377,12 @@ public class DefaultClientService implements ClientService {
         if (user == null) {
             throw new NotFoundException();
         }
+        
+        Customer customer = customerDao.getCustomerByCustomerId(clientGroup.getCustomerId());
+        
+        if (customer == null) {
+            throw new NotFoundException();
+        }
 
         try {
             clientDao.removeUserFromGroup(user.getUniqueId(), clientGroup);
